requirements.txt
Django==1.10.3
django-allauth==0.28.0
django-cors-headers==1.3.1
django-rest-auth==0.8.2
djangorestframework==3.5.3
djangorestframework-jwt==1.8.0
oauthlib==2.0.0
psycopg2==2.6.2
PyJWT==1.4.2
python-openid==2.2.5
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
